LibCompress: Rename LZWDecoder.h => Lzw.h
