+++ Userland/Libraries/LibCompress/Lzw.h
